Arreglo Herramientas y Meteriales todo ok en mi pc

+++ src/Herramientas/Desgastador/factorDesgastador.java
@@ -2,5 +2,5 @@ package Herramientas.Desgastador;
 
 public interface factorDesgastador {
 
-    int calcularDesgaste();
+    float calcularDesgaste();
 }
